feat: 设置页使用 tab 组织，初始化项目移入独立 tab
